NIAD-1141 Spine response handling extracted; additional logging
